renames extensionPointTestUtils -> starterBrickTestUtils
