+++ src/starterBricks/contextMenu.test.ts
@@ -21,7 +21,7 @@ import { type StarterBrickConfig } from "@/starterBricks/types";
 import { validateRegistryId } from "@/types/helpers";
 import { type Metadata } from "@/types/registryTypes";
 import { type BrickPipeline } from "@/bricks/types";
-import { RootReader } from "@/starterBricks/extensionPointTestUtils";
+import { RootReader } from "@/starterBricks/starterBrickTestUtils";
 import blockRegistry from "@/bricks/registry";
 import {
   type ContextMenuConfig,
