+++ src/starterBricks/helpers.test.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { getDocument, tick } from "@/starterBricks/extensionPointTestUtils";
+import { getDocument, tick } from "@/starterBricks/starterBrickTestUtils";
 import { awaitElementOnce } from "@/starterBricks/helpers";
 import { ensureMocksReset, requestIdleCallback } from "@shopify/jest-dom-mocks";
